fix @ProtonMail captcha endpoint/iframe URL resolving logic, #443
All the available API entry points are https-based, so it's ok if we just hardcode the "https" scheme in the patches. The @ProtonMail  captcha endpoint at the moment is only available for the https://mail.protonmail.com API entry point only though see #419.

List of the available API entry points:
  - https://mail.protonmail.com
  - https://app.protonmail.ch
  - https://protonmailrmez3lotccipshtkleegetolb73fuirgj7r4o4vfu7ozyd.onion